🎉 init(repo): begin a library for `nongli`
`Nongli` is a Chinese traditional lunisolar calendar, used almost as
widely as Gregorian calendar. However, I failed to find a rather neat
implementation of calculations of `nongli`. As is known, `nongli` is
based on the movements of the Sun-Earth-Moon system, which depends on
astronomical observations more than mathematical formulations.

Luckily, Mr. Xu J.W. developed a website called `SXWNL` for `nongli`
data from 4713 B.C. to 9999 A.D. in JavaScript. But seeing my poor
knowledge in astronomy, I am not able to refactor Xu's code in C++.
Nor can I develop a new version of my own. So it might be a best idea
for me to pre-calculate all data using Xu's code via a Python spider,
and develop a C/C++ library of my own based on the data.

Actually, `SXWNL` does have a C++ version, but it is too difficult to
read and maintain. I have to admit that I am not sure at all if the
data calculated with `SXWNL` are all correct. But I am sure that the
fitting methods I use work with the corrected data as well as those I
got. If I am still alive then, I promise to modify and correct them.

See: https://github.com/sxwnl/sxwnl-cpp