🌐 feat(data): parse `hanzi` in raw data text
As a significant part of Chinese culture, related info is certainly
expressed with `hanzi`, but I never know with which encoding others
would open my source files. All files are saved with UTF-8 and thus are
supposed to be opened with UTF-8, too. For the sake of fool-proofing,
all the source files except for this one contain no non-ASCII chars.

Please DO open `hanzi.py` with UTF-8!