🐛 fix(data): update `cyue` and `cjie` reasonably
The `c` prefix in `cyue` and `cjie` means cumulative. Both `cyue` and
`cjie` counts from the leading unit (`yue` or `jieqi`) of the minimal
`nian` (for lunar `cyue`) or `sui` (for solar `cjie`), instead of the
leading item fetched from remote.

Here a `sui` starts with a `dongzhi` incl. and ends with another one
excl., while a `nian` starts with a `p01` (i.e. `zhengyue`) and ends
with another one excl., where `p` is short for `ping` in Chinese. For
example, `p01` of the minimal `nian` is `cyue` 0, and `lichun` of the
next-minimal `sui` is `cjie` 27.