🗑️ refactor(data): allow right shift on negatives
In Python and since C++ >= 20, right shift on negatives, a >> b (a < 0),
performs an arithmetic right shift, while right shift on negatives until
C++ <= 17 is implementation-defined. Also, most compilers perform arith-
metically, and the others are taken out of account here. Thus the code
for translating the right shift operands to positives is removed.