✨ feat(data): export `nongli` data to C++ code
By specifying `MIN` and `MAX` in Section 0 of `coefs.ipynb`, data from
`MIN` incl. to `MAX` incl. are fitted, generated and exported to source
`fit/data.cpp`. It is OK if `MIN` = -5000 and `MAX` = 9999, but I do not
want Git to maintain such a huge source code file, so I pre-exported the
data from 2024 incl. to 2029 incl., and restored `MIN` and `MAX`.