✨ feat(fit): complete fittings for various data
It is worth mentioning that `std::sin` in `<cmath>` is not `constexpr`,
so I have implemented a set of such trigonometric functions of my own.
Trig functions, especially `sin` and `tan`, are used to estimate EoT
(equation of time) contributed by eccentricity of the Earth's orbit and
the obliquity of the ecliptic. `Bazi` requires real solar time, which is
determined by the Sun's apparent location, instead of UTC time.