✨ feat(fit): convert between `ryue` and `nyue`
Here `ryue` is a manually specified 8-bit integer. The lowest bit of it
is set 1 on a `runyue` and set 0 otherwise, and the rest of bits mean
the `yue` ordinal counting from 1. For example, the `ryue` of `zhengyue`
(p01, leading `yue` of a `nian`) is 2 = (1 << 1) | 0, and the `ryue` of
`run eryue` (r02) is 5 = (2 << 1) | 1. And `nyue` counts from 0 how many
`yue`s has passed after `zhengyue` (p01).