✨ feat(fit): add `next_` functions for `riqi`
The function `next_cyue` returns the same `tian` after `step` `cyue`s.
If the given `rizi` is 30th of the `yue` and the destination `yue` has
29 days only, then the 29th is returned. For example, you may get 2025-
p02-29 by moving 1 `cyue` forward 2025-p01-30.

Similarly, `next_nian` returns the same `ryue` and `tian` after `step`
`nian`s, and serves as a calculator of `shengri` (birthday in `nongli`).
Same as `next_cyue`, if you move a 30th to a 29-day `yue`, the 29th will
be returned. Thanks to Ms. Zhao Y.H., I've made it clear when one born
in `runyue` is accustomed to `guo shengri` (celebrate the birthday).
If (s)he was born in `r02`, (s)he celebrates it on the same `tian` in
`r02` as well, but when `r02` does not occur in a certain `nian`, (s)he
celebrates it on the same `tian` in `p02` instead.

Before I asked Zhao, I had searched Zhihu and Xiaohongshu for hints, but
it did frustrate me I failed to find any, including some advising others
to celebrate in Gregorian in this case. Many thanks to Zhao Y.H. again.