⚡️ perf(fit): improve perf by reducing precisions
It is enough for `cos` and `sin` to keep a precision of 1p-30, about
1e-9. To cut off unnecessary floating-point multiplications, `ccos` and
`csin` take the places of `cos` and `sin`, respectively, where ccos(x)
= cos(2 PI x) and csin(x) similar. Meanwhile, the callers of `ccos` and
`csin` prepare angles in cycle instead of radian. And as for fittings of
`oecc` (eccentricity of the Earth's orbit) and `eobl` (obliquity of the
ecliptic), quadratic terms and higher are neglected.