🔧 fix(data): use conda-forge channel only
Amazingly Anaconda requires a charged license for commercial usage of
enterprises with 200+ people. But commercial fees do not apply to conda-
forge channel, which is maintained by the community. That is why a new
conda environment is created with all packages from conda-forge.

See: https://www.anaconda.com/pricing/terms-of-service-faqs