⚡️ perf(math): use fixed-point numbers
As is widely known, integral operations are much faster than floating-
point operations on the CPU, and so is fixed-point operations based on
integral ones. However, different from industrial conventions, fixed-
point operations are not saturated on overflows and underflows. And the
trig functions `cosq` and `sinq` give an absolute error within 2 units,
i.e. 2 ** -31, which is surely satisfactory here. By experiments with
GCC 14 and -O2, `cosq` is 2x as fast as `std::cos<double>`.

I've made it!!